= Version 1.3.5.8 =
Améliorations

* ST646 - Envoi differé du mail de confirmation de commande (Possibilité d'envoyer l'email après avoir reçu la totalité du paiement de la commande)
* ST647 - Administration commande Box Commentaire du client


Corrections

* ST641 - Frais de livraison et Taxes sur PDF des Devis.
* ST642 - Pagination des produits ( Résultat de recherche)
* ST643 - Coupon de réduction
* ST644 - Duplication  de produit (Problème avec Yoast SEO Plugin (lien vers le produit dupliqué ->Array ))
* ST645 - Affichage d'une nouvelle adresse
* ST648 - Options obligatoires (Les options de produits étaient constamment obligatoires : problème corrigé.)

git-svn-id: https://plugins.svn.wordpress.org/wpshop/trunk@748679 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/librairies/purchase/coupons.class.php
@@ -143,10 +143,11 @@ class wpshop_coupons	{
 	*/
 	function save_coupon_custom_informations()
 	{
+		
 		if( !empty($_REQUEST['post_ID']) && (get_post_type($_REQUEST['post_ID']) == WPSHOP_NEWTYPE_IDENTIFIER_COUPON) )
 		{
 			update_post_meta($_REQUEST['post_ID'], 'wpshop_coupon_code', $_REQUEST['coupon_code']);
-			update_post_meta($_REQUEST['post_ID'], 'wpshop_coupon_discount_value', intval($_REQUEST['coupon_discount_amount']));
+			update_post_meta($_REQUEST['post_ID'], 'wpshop_coupon_discount_value', floatval($_REQUEST['coupon_discount_amount']) );
 			update_post_meta($_REQUEST['post_ID'], 'wpshop_coupon_discount_type', $_REQUEST['coupon_type']);
 			update_post_meta($_REQUEST['post_ID'], 'wpshop_coupon_individual_use', '');
 			update_post_meta($_REQUEST['post_ID'], 'wpshop_coupon_product_ids', '');
